+++ AntonBabunin/src/test/java/tests/RunTestSalesman.java
